+++ src/modules/file/components/index.ts
@@ -1,0 +1,3 @@
+export * from './file/file.component';
+export * from './fileInput/fileInput.component';
+export * from './uploadButton/uploadButton.component';