+++ src/modules/file/directive/index.ts
@@ -1,0 +1,3 @@
+export * from './fileInput/fileInput.directive';
+export * from './fileTemplate/fileTemplate.context';
+export * from './fileTemplate/fileTemplate.directive';