+++ src/modules/file/index.ts
@@ -1,0 +1,5 @@
+export * from './components';
+export * from './directive';
+export * from './interfaces';
+export * from './misc/tokens';
+export * from './modules';