+++ src/modules/file/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './fileVisual/fileVisual.interface';