+++ src/pipes/firstUppercase/firstUppercase.pipe.ts
@@ -14,13 +14,13 @@ export class FirstUppercaseSAPipe implements PipeTransform
      * Converts first letter of text to uppercase
      * @param value - Value to be converted
      */
-    public transform(value: string|undefined|null): string|undefined|null 
+public transform<TValue extends string|undefined|null>(value: TValue): TValue 
     {
         if(!value)
         {
             return value;
         }
 
-        return value[0].toUpperCase() + value.substring(1);
+        return value[0].toUpperCase() + value.substring(1) as TValue;
     }
 }